Refactor HenkiloService into HenkiloService and HenkiloModificationService.
