+++ oppijanumerorekisteri-service/src/main/java/fi/vm/sade/oppijanumerorekisteri/services/HenkiloModificationService.java
@@ -1,0 +1,28 @@
+package fi.vm.sade.oppijanumerorekisteri.services;
+
+import fi.vm.sade.oppijanumerorekisteri.dto.*;
+import fi.vm.sade.oppijanumerorekisteri.models.Henkilo;
+
+import java.util.List;
+
+public interface HenkiloModificationService {
+    HenkiloUpdateDto updateHenkilo(HenkiloUpdateDto henkiloUpdateDto);
+
+    HenkiloReadDto forceUpdateHenkilo(HenkiloForceUpdateDto henkiloUpdateDto);
+
+    Henkilo update(Henkilo henkilo);
+
+    Henkilo disableHenkilo(String oid);
+
+    FindOrCreateWrapper<HenkiloPerustietoDto> findOrCreateHenkiloFromPerustietoDto(HenkiloPerustietoDto henkiloPerustietoDto);
+
+    List<HenkiloPerustietoDto> findOrCreateHenkiloFromPerustietoDto(List<HenkiloPerustietoDto> henkilot);
+
+    HenkiloDto createHenkilo(HenkiloCreateDto henkiloDto);
+
+    Henkilo createHenkilo(Henkilo henkiloCreate);
+
+    Henkilo createHenkilo(Henkilo henkiloCreate, String kasittelijaOid, boolean validate);
+
+
+}
