+++ oppijanumerorekisteri-service/src/test/java/fi/vm/sade/oppijanumerorekisteri/services/HenkiloServiceTest.java
@@ -1,6 +1,5 @@
 package fi.vm.sade.oppijanumerorekisteri.services;
 
-import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
 import com.querydsl.core.types.Predicate;
 import fi.vm.sade.kayttooikeus.dto.KayttooikeudetDto;
@@ -13,8 +12,10 @@ import fi.vm.sade.oppijanumerorekisteri.exceptions.NotFoundException;
 import fi.vm.sade.oppijanumerorekisteri.mappers.EntityUtils;
 import fi.vm.sade.oppijanumerorekisteri.mappers.OrikaConfiguration;
 import fi.vm.sade.oppijanumerorekisteri.models.Henkilo;
-import fi.vm.sade.oppijanumerorekisteri.models.YhteystiedotRyhma;
-import fi.vm.sade.oppijanumerorekisteri.repositories.*;
+import fi.vm.sade.oppijanumerorekisteri.repositories.HenkiloRepository;
+import fi.vm.sade.oppijanumerorekisteri.repositories.HenkiloViiteRepository;
+import fi.vm.sade.oppijanumerorekisteri.repositories.KansalaisuusRepository;
+import fi.vm.sade.oppijanumerorekisteri.repositories.KielisyysRepository;
 import fi.vm.sade.oppijanumerorekisteri.repositories.criteria.HenkiloCriteria;
 import fi.vm.sade.oppijanumerorekisteri.repositories.criteria.YhteystietoCriteria;
 import fi.vm.sade.oppijanumerorekisteri.repositories.dto.YhteystietoHakuDto;
@@ -26,15 +27,19 @@ import fi.vm.sade.oppijanumerorekisteri.validators.HenkiloUpdatePostValidator;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.*;
+import org.mockito.ArgumentCaptor;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.Spy;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.util.ReflectionTestUtils;
 
-import java.time.LocalDate;
-import java.time.Month;
-import java.util.*;
+import java.util.Collections;
+import java.util.Date;
+import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -43,8 +48,6 @@ import static java.util.Arrays.asList;
 import static java.util.Collections.*;
 import static java.util.stream.Collectors.toSet;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.tuple;
-import static org.mockito.AdditionalAnswers.returnsFirstArg;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.*;
 
@@ -87,7 +90,6 @@ public class HenkiloServiceTest {
     public void setup() {
         ReflectionTestUtils.setField(this.service, "mapper", this.mapper);
         ReflectionTestUtils.setField(this.service, "yhteystietoConverter", new YhteystietoConverter());
-        ReflectionTestUtils.setField(this.service, "oidGenerator", new MockOidGenerator());
     }
 
     @Test
@@ -274,12 +276,6 @@ public class HenkiloServiceTest {
     }
 
     @Test
-    public void generateOid() {
-        String oid = ReflectionTestUtils.invokeMethod(service, "getFreePersonOid");
-        assertThat(oid).isNotNull();
-    }
-
-    @Test
     public void getHenkiloPerustietoByOidsTest() {
         Date modified = new Date();
         HenkiloPerustietoDto henkiloMock = DtoUtils.createHenkiloPerustietoDto("arpa", "arpa", "kuutio",
@@ -336,60 +332,6 @@ public class HenkiloServiceTest {
         assertThat(henkiloTyypit.get(0)).isEqualTo("VIRKAILIJA");
     }
 
-    @Test
-    public void updateHenkiloFromHenkiloUpdateDto() {
-        Henkilo henkilo = EntityUtils.createHenkilo("arpa noppa", "arpa", "kuutio", "123456-9999",
-                "1.2.3.4.5", false, "fi", "suomi", "246",
-                new Date(), new Date(), "1.2.3.4.1", "arpa@kuutio.fi");
-        long yhteystiedotRyhmaId = henkilo.getYhteystiedotRyhma().stream().mapToLong(YhteystiedotRyhma::getId).max().orElse(0L) + 1;
-        henkilo.getYhteystiedotRyhma().add(YhteystiedotRyhma.builder()
-                .ryhmaKuvaus("yhteystietotyyppi5").ryhmaAlkuperaTieto("alkupera1")
-                .id(yhteystiedotRyhmaId).readOnly(true)
-                .build());
-        HenkiloUpdateDto henkiloUpdateDto = DtoUtils.createHenkiloUpdateDto("arpa", "arpa", "kuutio",
-                "123456-9999", "1.2.3.4.5", "sv", "svenska", "246",
-                "arpa@kuutio.fi");
-        henkiloUpdateDto.getYhteystiedotRyhma().add(YhteystiedotRyhmaDto.builder()
-                .id(yhteystiedotRyhmaId).readOnly(false)
-                .ryhmaKuvaus("readonly").ryhmaAlkuperaTieto("readonly")
-                .build());
-        ArgumentCaptor<Henkilo> argument = ArgumentCaptor.forClass(Henkilo.class);
-        given(this.henkiloDataRepositoryMock.findByOidHenkiloIsIn(Collections.singletonList(henkiloUpdateDto.getOidHenkilo())))
-        .willReturn(Collections.singletonList(henkilo));
-        given(userDetailsHelperMock.getCurrentUserOid()).willReturn("1.2.3.4.1");
-        given(this.kielisyysRepositoryMock.findByKieliKoodi(anyString()))
-                .willReturn(Optional.of(EntityUtils.createKielisyys("fi", "suomi")));
-        given(this.kansalaisuusRepositoryMock.findOrCreate(anyString()))
-                .willReturn(EntityUtils.createKansalaisuus("246"));
-        given(henkiloDataRepositoryMock.save(any(Henkilo.class)))
-                .willAnswer(returnsFirstArg());
-
-        this.service.updateHenkilo(henkiloUpdateDto);
-        verify(this.henkiloDataRepositoryMock).save(argument.capture());
-        verify(this.kayttooikeusClient, times(1)).ldapSynkroniseHenkilo(eq("1.2.3.4.5"));
-
-        assertThat(argument.getValue().getAidinkieli().getKieliKoodi()).isEqualTo("fi");
-        assertThat(argument.getValue().getAidinkieli().getKieliTyyppi()).isEqualTo("suomi");
-        assertThat(argument.getValue().getAsiointiKieli().getKieliKoodi()).isEqualTo("fi");
-        assertThat(argument.getValue().getAsiointiKieli().getKieliTyyppi()).isEqualTo("suomi");
-
-        assertThat(argument.getValue().getKielisyys().size()).isEqualTo(1);
-        assertThat(argument.getValue().getKielisyys().iterator().next().getKieliKoodi()).isEqualTo("fi");
-        assertThat(argument.getValue().getKielisyys().iterator().next().getKieliTyyppi()).isEqualTo("suomi");
-
-        assertThat(argument.getValue().getKansalaisuus().size()).isEqualTo(1);
-        assertThat(argument.getValue().getKansalaisuus().iterator().next().getKansalaisuusKoodi()).isEqualTo("246");
-
-        assertThat(argument.getValue().getYhteystiedotRyhma())
-                .extracting("id", "ryhmaAlkuperaTieto", "ryhmaKuvaus")
-                .containsExactlyInAnyOrder(
-                        tuple(null, "alkupera2", "yhteystietotyyppi7"),
-                        tuple(yhteystiedotRyhmaId, "alkupera1", "yhteystietotyyppi5"));
-
-        assertThat(argument.getValue().getYhteystiedotRyhma())
-                .flatExtracting("yhteystieto").extracting("yhteystietoTyyppi", "yhteystietoArvo")
-                .containsExactly(tuple(YhteystietoTyyppi.YHTEYSTIETO_MATKAPUHELINNUMERO, "arpa@kuutio.fi"));
-    }
 
     @Test
     public void findHenkiloViitteesTest() {
@@ -413,221 +355,4 @@ public class HenkiloServiceTest {
         results = this.service.findHenkiloViittees(criteria);
         assertThat(results).size().isEqualTo(2);
     }
-
-    @Test
-    public void findOrCreateHenkiloFromPerustietoDto() {
-        HenkiloPerustietoDto henkiloPerustietoDtoMock = DtoUtils.createHenkiloPerustietoDto("arpa", "arpa", "kuutio",
-                "123456-9999", null, "fi", "suomi", "246", emptyList(), emptyList(), null, new Date());
-
-        given(this.henkiloDataRepositoryMock.findByHetu(henkiloPerustietoDtoMock.getHetu())).willReturn(Optional.empty());
-        doAnswer(returnsFirstArg()).when(this.service).createHenkilo(any(Henkilo.class));
-        assertThat(this.service.findOrCreateHenkiloFromPerustietoDto(henkiloPerustietoDtoMock).getDto())
-                .isEqualToComparingFieldByFieldRecursively(henkiloPerustietoDtoMock);
-    }
-
-    @Test
-    public void findOrCreateHenkiloFromPerustietoDtoHenkiloFoundByOid() {
-        LocalDate syntymaaika = LocalDate.now();
-        Date modified = new Date();
-        HenkiloPerustietoDto henkiloPerustietoDtoMock = DtoUtils.createHenkiloPerustietoDto("arpa", "arpa", "kuutio",
-                "123456-9999", "1.2.3.4.5", "fi", "suomi", "246", null, null, syntymaaika, modified);
-        Henkilo henkilo = EntityUtils.createHenkilo("arpa", "arpa", "kuutio", "123456-9999", "1.2.3.4.5", false,
-                "fi", "suomi", "246", modified, modified, "1.2.3.4.1", "arpa@kuutio.fi", syntymaaika);
-
-        given(this.henkiloDataRepositoryMock.findByOidHenkilo(any()))
-                .willReturn(Optional.of(henkilo));
-        assertThat(this.service.findOrCreateHenkiloFromPerustietoDto(henkiloPerustietoDtoMock).getDto())
-                .isEqualToComparingFieldByFieldRecursively(henkiloPerustietoDtoMock);
-        verify(this.henkiloDataRepositoryMock).findByOidHenkilo(eq(henkiloPerustietoDtoMock.getOidHenkilo()));
-    }
-
-    @Test(expected = NotFoundException.class)
-    public void findOrCreateHenkiloFromPerustietoDtoHenkiloNotFoundByOid() {
-        HenkiloPerustietoDto henkiloPerustietoDtoMock = DtoUtils.createHenkiloPerustietoDto("arpa", "arpa", "kuutio",
-                "123456-9999", "1.2.3.4.5", "fi", "suomi", "246", null, null, null, new Date());
-        Henkilo henkilo = EntityUtils.createHenkilo("arpa", "arpa", "kuutio", "123456-9999", "1.2.3.4.5", false,
-                "fi", "suomi", "246", new Date(), new Date(), "1.2.3.4.1", "arpa@kuutio.fi");
-
-        given(this.henkiloDataRepositoryMock.findByOidHenkilo(any()))
-                .willReturn(Optional.empty());
-        this.service.findOrCreateHenkiloFromPerustietoDto(henkiloPerustietoDtoMock);
-        verify(this.henkiloDataRepositoryMock).findByOidHenkilo(eq(henkiloPerustietoDtoMock.getOidHenkilo()));
-    }
-
-    @Test
-    public void findOrCreateHenkiloFromPerustietoDtoHenkiloFoundByHetu() {
-        LocalDate syntymaaika = LocalDate.now();
-        Date modified = new Date();
-        HenkiloPerustietoDto henkiloPerustietoDtoInput = DtoUtils.createHenkiloPerustietoDto(null, null, null,
-                "123456-9999", null, null, null, null, null, null, syntymaaika, modified);
-        HenkiloPerustietoDto henkiloPerustietoDtoMock = DtoUtils.createHenkiloPerustietoDto("arpa", "arpa", "kuutio",
-                "123456-9999", "", "fi", "suomi", "246", null, null, syntymaaika, modified);
-        Henkilo henkilo = EntityUtils.createHenkilo("arpa", "arpa", "kuutio", "123456-9999", "", false,
-                "fi", "suomi", "246", modified, new Date(), "1.2.3.4.1", "arpa@kuutio.fi", syntymaaika);
-
-
-        given(this.henkiloDataRepositoryMock.findByHetu(henkiloPerustietoDtoInput.getHetu())).willReturn(Optional.of(henkilo));
-        assertThat(this.service.findOrCreateHenkiloFromPerustietoDto(henkiloPerustietoDtoInput).getDto())
-                .isEqualToComparingFieldByFieldRecursively(henkiloPerustietoDtoMock);
-    }
-
-    @Test
-    public void createHenkiloShouldSetEmptyHetuToNull() {
-        when(henkiloDataRepositoryMock.save(any(Henkilo.class)))
-                .thenAnswer(returnsFirstArg());
-        HenkiloCreateDto input = new HenkiloCreateDto();
-        input.setHetu("");
-
-        HenkiloDto output = service.createHenkilo(input);
-
-        assertThat(output.getHetu()).isNull();
-        ArgumentCaptor<Henkilo> argumentCaptor = ArgumentCaptor.forClass(Henkilo.class);
-        verify(henkiloDataRepositoryMock).save(argumentCaptor.capture());
-        Henkilo saved = argumentCaptor.getValue();
-        assertThat(saved.getHetu()).isNull();
-    }
-
-    @Test
-    public void createHenkiloShouldSaveHetu() {
-        when(henkiloDataRepositoryMock.save(any(Henkilo.class)))
-                .thenAnswer(returnsFirstArg());
-        HenkiloCreateDto input = new HenkiloCreateDto();
-        input.setHetu("310817A983J");
-        input.setSyntymaaika(LocalDate.of(2017, Month.OCTOBER, 6));
-        input.setSukunimi("2");
-
-        HenkiloDto output = service.createHenkilo(input);
-
-        assertThat(output.getHetu()).isEqualTo("310817A983J");
-        ArgumentCaptor<Henkilo> argumentCaptor = ArgumentCaptor.forClass(Henkilo.class);
-        verify(henkiloDataRepositoryMock).save(argumentCaptor.capture());
-        Henkilo saved = argumentCaptor.getValue();
-        assertThat(saved.getHetu()).isEqualTo("310817A983J");
-        assertThat(saved.getSyntymaaika()).isEqualTo("2017-08-31");
-        assertThat(saved.getSukupuoli()).isEqualTo("1");
-    }
-
-    @Test
-    public void createHenkiloShouldSaveSyntymaaika() {
-        when(henkiloDataRepositoryMock.save(any(Henkilo.class)))
-                .thenAnswer(returnsFirstArg());
-        HenkiloCreateDto input = new HenkiloCreateDto();
-        input.setSyntymaaika(LocalDate.of(2017, Month.OCTOBER, 4));
-
-        HenkiloDto output = service.createHenkilo(input);
-
-        assertThat(output.getSyntymaaika()).isEqualTo("2017-10-04");
-        ArgumentCaptor<Henkilo> argumentCaptor = ArgumentCaptor.forClass(Henkilo.class);
-        verify(henkiloDataRepositoryMock).save(argumentCaptor.capture());
-        Henkilo saved = argumentCaptor.getValue();
-        assertThat(saved.getSyntymaaika()).isEqualTo("2017-10-04");
-    }
-
-    @Test
-    public void createHenkiloShouldSaveSukupuoli() {
-        when(henkiloDataRepositoryMock.save(any(Henkilo.class)))
-                .thenAnswer(returnsFirstArg());
-        HenkiloCreateDto input = new HenkiloCreateDto();
-        input.setSukupuoli("1");
-
-        HenkiloDto output = service.createHenkilo(input);
-
-        assertThat(output.getSukupuoli()).isEqualTo("1");
-        ArgumentCaptor<Henkilo> argumentCaptor = ArgumentCaptor.forClass(Henkilo.class);
-        verify(henkiloDataRepositoryMock).save(argumentCaptor.capture());
-        Henkilo saved = argumentCaptor.getValue();
-        assertThat(saved.getSukupuoli()).isEqualTo("1");
-    }
-
-    @Test
-    public void updateHenkiloShouldSetEmptyHetuToNull() {
-        when(henkiloDataRepositoryMock.findByOidHenkiloIsIn(any()))
-                .thenReturn(Lists.newArrayList(new Henkilo()));
-        when(henkiloDataRepositoryMock.save(any(Henkilo.class)))
-                .thenAnswer(returnsFirstArg());
-        HenkiloUpdateDto input = new HenkiloUpdateDto();
-        input.setHetu("");
-
-        HenkiloUpdateDto output = service.updateHenkilo(input);
-
-        assertThat(output.getHetu()).isNull();
-        ArgumentCaptor<Henkilo> argumentCaptor = ArgumentCaptor.forClass(Henkilo.class);
-        verify(henkiloDataRepositoryMock).save(argumentCaptor.capture());
-        Henkilo saved = argumentCaptor.getValue();
-        assertThat(saved.getHetu()).isNull();
-    }
-
-    @Test
-    public void updateHenkiloShouldSaveHetu() {
-        when(henkiloDataRepositoryMock.findByOidHenkiloIsIn(any()))
-                .thenReturn(Lists.newArrayList(new Henkilo()));
-        when(henkiloDataRepositoryMock.save(any(Henkilo.class)))
-                .thenAnswer(returnsFirstArg());
-        HenkiloUpdateDto input = new HenkiloUpdateDto();
-        input.setHetu("310817A983J");
-        input.setSyntymaaika(LocalDate.of(2017, Month.OCTOBER, 6));
-        input.setSukunimi("2");
-
-        HenkiloUpdateDto output = service.updateHenkilo(input);
-
-        assertThat(output.getHetu()).isEqualTo("310817A983J");
-        ArgumentCaptor<Henkilo> argumentCaptor = ArgumentCaptor.forClass(Henkilo.class);
-        verify(henkiloDataRepositoryMock).save(argumentCaptor.capture());
-        Henkilo saved = argumentCaptor.getValue();
-        assertThat(saved.getHetu()).isEqualTo("310817A983J");
-        assertThat(saved.getSyntymaaika()).isEqualTo("2017-08-31");
-        assertThat(saved.getSukupuoli()).isEqualTo("1");
-    }
-
-    @Test
-    public void updateHenkiloShouldSaveSyntymaaika() {
-        when(henkiloDataRepositoryMock.findByOidHenkiloIsIn(any()))
-                .thenReturn(Lists.newArrayList(new Henkilo()));
-        when(henkiloDataRepositoryMock.save(any(Henkilo.class)))
-                .thenAnswer(returnsFirstArg());
-        HenkiloUpdateDto input = new HenkiloUpdateDto();
-        input.setSyntymaaika(LocalDate.of(2017, Month.OCTOBER, 4));
-
-        HenkiloUpdateDto output = service.updateHenkilo(input);
-
-        assertThat(output.getSyntymaaika()).isEqualTo("2017-10-04");
-        ArgumentCaptor<Henkilo> argumentCaptor = ArgumentCaptor.forClass(Henkilo.class);
-        verify(henkiloDataRepositoryMock).save(argumentCaptor.capture());
-        Henkilo saved = argumentCaptor.getValue();
-        assertThat(saved.getSyntymaaika()).isEqualTo("2017-10-04");
-    }
-
-    @Test
-    public void updateHenkiloShouldSaveSukupuoli() {
-        when(henkiloDataRepositoryMock.findByOidHenkiloIsIn(any()))
-                .thenReturn(Lists.newArrayList(new Henkilo()));
-        when(henkiloDataRepositoryMock.save(any(Henkilo.class)))
-                .thenAnswer(returnsFirstArg());
-        HenkiloUpdateDto input = new HenkiloUpdateDto();
-        input.setSukupuoli("1");
-
-        HenkiloUpdateDto output = service.updateHenkilo(input);
-
-        assertThat(output.getSukupuoli()).isEqualTo("1");
-        ArgumentCaptor<Henkilo> argumentCaptor = ArgumentCaptor.forClass(Henkilo.class);
-        verify(henkiloDataRepositoryMock).save(argumentCaptor.capture());
-        Henkilo saved = argumentCaptor.getValue();
-        assertThat(saved.getSukupuoli()).isEqualTo("1");
-    }
-
-    @Test
-    public void createShouldSkipEmptyPassinumerot() {
-        when(henkiloDataRepositoryMock.save(any(Henkilo.class)))
-                .thenAnswer(returnsFirstArg());
-        HenkiloCreateDto input = new HenkiloCreateDto();
-        input.setPassinumerot(Stream.of("passi123", null, " ").collect(toSet()));
-
-        HenkiloDto output = service.createHenkilo(input);
-
-        ArgumentCaptor<Henkilo> argumentCaptor = ArgumentCaptor.forClass(Henkilo.class);
-        verify(henkiloDataRepositoryMock).save(argumentCaptor.capture());
-        Henkilo saved = argumentCaptor.getValue();
-        assertThat(saved.getPassinumerot()).containsExactly("passi123");
-    }
-
 }
