+++ oppijanumerorekisteri-service/src/test/java/fi/vm/sade/oppijanumerorekisteri/services/impl/HenkiloServiceImplTest.java
@@ -203,117 +203,4 @@ public class HenkiloServiceImplTest {
         assertThat(result).hasSize(0);
     }
 
-    @Test
-    public void findOrCreateHenkiloFromPerustietoDtoShouldFindByOid() {
-        HenkiloPerustietoDto input = HenkiloPerustietoDto.builder().oidHenkilo("oid1").build();
-        Henkilo henkilo = new Henkilo();
-        when(henkiloRepository.findByOidHenkilo(any())).thenReturn(Optional.of(henkilo));
-        when(orikaConfiguration.map(any(), eq(HenkiloPerustietoDto.class))).thenReturn(new HenkiloPerustietoDto());
-
-        FindOrCreateWrapper<HenkiloPerustietoDto> output = impl.findOrCreateHenkiloFromPerustietoDto(input);
-
-        verify(henkiloRepository).findByOidHenkilo(eq("oid1"));
-        verify(orikaConfiguration).map(eq(henkilo), eq(HenkiloPerustietoDto.class));
-        verify(henkiloRepository, never()).save(any(Henkilo.class));
-    }
-
-    @Test
-    public void findOrCreateHenkiloFromPerustietoDtoShouldThrowByOid() {
-        HenkiloPerustietoDto input = HenkiloPerustietoDto.builder().oidHenkilo("oid1").build();
-        Henkilo henkilo = new Henkilo();
-        when(henkiloRepository.findByOidHenkilo(any())).thenReturn(Optional.empty());
-
-        Throwable throwable = catchThrowable(() -> impl.findOrCreateHenkiloFromPerustietoDto(input));
-
-        assertThat(throwable).isInstanceOf(NotFoundException.class);
-        verify(henkiloRepository).findByOidHenkilo(eq("oid1"));
-        verify(henkiloRepository, never()).save(any(Henkilo.class));
-    }
-
-    @Test
-    public void findOrCreateHenkiloFromPerustietoDtoShouldFindByExternalId() {
-        HenkiloPerustietoDto input = HenkiloPerustietoDto.builder().externalIds(singletonList("externalid1")).build();
-        Henkilo henkilo = new Henkilo();
-        when(henkiloRepository.findByExternalIds(any())).thenReturn(singleton(henkilo));
-        when(orikaConfiguration.map(any(), eq(HenkiloPerustietoDto.class))).thenReturn(new HenkiloPerustietoDto());
-
-        FindOrCreateWrapper<HenkiloPerustietoDto> output = impl.findOrCreateHenkiloFromPerustietoDto(input);
-
-        verify(henkiloRepository).findByExternalIds(eq(singletonList("externalid1")));
-        verify(orikaConfiguration).map(eq(henkilo), eq(HenkiloPerustietoDto.class));
-        verify(henkiloRepository, never()).save(any(Henkilo.class));
-    }
-
-    @Test
-    public void findOrCreateHenkiloFromPerustietoDtoShouldFindByIdentification() {
-        IdentificationDto identification = new IdentificationDto();
-        identification.setIdpEntityId("key");
-        identification.setIdentifier("value1");
-        HenkiloPerustietoDto input = HenkiloPerustietoDto.builder().identifications(singletonList(identification)).build();
-        Henkilo henkilo = new Henkilo();
-        when(henkiloRepository.findByIdentifications(any())).thenReturn(singleton(henkilo));
-        when(orikaConfiguration.map(any(), eq(HenkiloPerustietoDto.class))).thenReturn(new HenkiloPerustietoDto());
-
-        FindOrCreateWrapper<HenkiloPerustietoDto> output = impl.findOrCreateHenkiloFromPerustietoDto(input);
-
-        verify(henkiloRepository).findByIdentifications(eq(singletonList(identification)));
-        verify(orikaConfiguration).map(eq(henkilo), eq(HenkiloPerustietoDto.class));
-        verify(henkiloRepository, never()).save(any(Henkilo.class));
-    }
-
-    @Test
-    public void createHenkiloShouldSetOppijanumero() {
-        when(oidGenerator.generateOID()).thenReturn("oid1");
-        when(henkiloRepository.save(any(Henkilo.class))).thenAnswer(returnsFirstArg());
-        Henkilo input = new Henkilo();
-        Henkilo output;
-
-        input.setYksiloity(false);
-        output = impl.createHenkilo(input);
-        assertThat(output)
-                .extracting(Henkilo::getOidHenkilo, Henkilo::getOppijanumero, Henkilo::isYksiloity)
-                .containsExactly("oid1", null, false);
-
-        input.setYksiloity(true);
-        output = impl.createHenkilo(input);
-        assertThat(output)
-                .extracting(Henkilo::getOidHenkilo, Henkilo::getOppijanumero, Henkilo::isYksiloity)
-                .containsExactly("oid1", "oid1", true);
-
-        input.setYksiloity(true);
-        input.setHetu("241197-9877");
-        Throwable throwable1 = catchThrowable(() -> impl.createHenkilo(input));
-        assertThat(throwable1)
-                .isInstanceOf(ValidationException.class)
-                .hasMessageContaining("Henkilöllä on hetu");
-
-        input.setYksiloity(true);
-        input.setHetu(null);
-        input.setDuplicate(true);
-        Throwable throwable2 = catchThrowable(() -> impl.createHenkilo(input));
-        assertThat(throwable2)
-                .isInstanceOf(ValidationException.class)
-                .hasMessageContaining("Henkilö on duplikaatti");
-    }
-
-    @Test
-    public void updateHenkiloShouldUpdateSlaves() {
-        when(userDetailsHelper.findCurrentUserOid()).thenReturn(Optional.of("käsittelijä"));
-        Henkilo input = new Henkilo();
-        input.setOidHenkilo("oid1");
-        when(henkiloRepository.save(any(Henkilo.class))).thenAnswer(returnsFirstArg());
-        Henkilo slave1 = new Henkilo();
-        Henkilo slave2 = new Henkilo();
-        when(henkiloRepository.findSlavesByMasterOid(any())).thenReturn(asList(slave1, slave2));
-
-        Henkilo output = impl.update(input);
-
-        ArgumentCaptor<Henkilo> henkiloCaptor = ArgumentCaptor.forClass(Henkilo.class);
-        verify(henkiloRepository).findSlavesByMasterOid(eq("oid1"));
-        verify(henkiloRepository, times(3)).save(henkiloCaptor.capture());
-        List<Henkilo> tallennetut = henkiloCaptor.getAllValues();
-        assertThat(tallennetut).extracting(Henkilo::getModified).isNotNull();
-        assertThat(tallennetut).extracting(Henkilo::getKasittelijaOid).containsOnly("käsittelijä");
-    }
-
 }
